Move SierraAccessConditions into a separate trait
